+++ src/assets/nodata-icon.tsx
@@ -2,13 +2,13 @@ import React from 'react';
 
 export const NoDataIcon: React.FC = () => {
   return (
-    <svg xmlns="http://www.w3.org/2000/svg" width={80} height={80} viewBox="-32 0 512 512">
+    <svg xmlns="http://www.w3.org/2000/svg" width={70} height={70} viewBox="-32 0 512 512">
       <path
-        fill="#8E939D"
+        fill="#A5AAAF"
         d="M189.14 512c-61.437 0-117.308-28.566-153.269-78.379a188.746 188.746 0 010-221.398L189.141 0l153.273 212.242c47.828 66.203 47.828 155.176 0 221.395C306.438 483.434 250.582 512 189.141 512zm0 0"
       ></path>
       <path
-        fill="#6B7280"
+        fill="#868993"
         d="M366.344 257.277c-10.27-2.312-20.91-3.636-31.875-3.636-80.14 0-145.328 65.203-145.328 145.328 0 24.011 6.054 46.55 16.437 66.496-.484-.938-1.062-1.824-1.531-2.778-4.957.372-9.883.872-14.922.872-61.441 0-117.309-28.567-153.27-78.38C16.93 358.99 5.63 329.212 1.672 298.696c-6.024 46.582 5.281 94.914 34.183 134.926C71.832 483.418 127.703 512 189.125 512c16.277 0 32.117-2.133 47.344-6.008a145.75 145.75 0 01-9.914-10.043c3.148 3.488 6.445 6.863 9.914 10.043 42.242-10.754 79.476-35.75 105.926-72.355 37.64-52.106 45.539-118.262 23.949-176.36zm0 0"
       ></path>
       <path
